Make example description and metadata general
The metadata dictionary as used in CTLearn has been eliminated and
replaced with an "example description", which is a list of dictionaries
that describe the example array at the corresponding index. These
contain the shape and dtype information needed to load the data. The
number of total and selected telescopes are available as class
attributes. To get the number of examples, a function is provided that
calculates the number of examples grouped by any parameter included in
the example. This general approach allows calculation of the number of
examples divided arbitrarily, for example, by gamma/hadron label, energy
bin, or multiplicity, as needed. However, as these correspond to the
selected events, getting the number of total pre-selection-cut examples
as needed to calculate q-factor will require the construction of a
second DL1DataReader.


Former-commit-id: 929d8a85c60f9ff591af6f9213ef7efb474c5205 [formerly 929d8a85c60f9ff591af6f9213ef7efb474c5205 [formerly 6d7e0a1cc914e2dba5c80f0126b42ec137fc95e8]]
Former-commit-id: 5a9953051e6eb734b51a7883e0880a9e67c18c26
Former-commit-id: fbf211033f6060e898565fa9f41cf72ba9ff1726